-태창금속 영업관리 목재단가(100%) 영업관리 제품품목정보관리(40%)

+++ src/main/java/com/mes/mesCrm/mesCrm/DTO/CRM_OUT_SUB.java
@@ -7,7 +7,7 @@ public class CRM_OUT_SUB {
     private String site_code;
     private String out_no;
     private String part_code;
-    private double qty;
+    private int qty;
     private String part_name;
     private String spec;
     private String unit_name;
@@ -17,8 +17,8 @@ public class CRM_OUT_SUB {
     private String supp_name;
     private String currency_type_name;
 
-    private double unit_price;
-    private double price;
+    private int unit_price;
+    private int price;
     private String update_date;
     private int rownum;
     private int rownum_page;
